requirements.txt
# Runtime dependencies (installed by Vercel during build)
chromadb>=0.5.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
